fix: remove wildcards from package-lock.json
With a checked in version of package-lock.json, running
an `npm install` would create a modified version of package-lock.json
with wilcard (') characters in the version strings of some packages.

I had to delete the existing package-lock.json and re-run `npm install`
to get a version of package-lock.json without wildcards in it.